* commit '0b9575d5bff84aa2ffa23dde037b19ec6c9a9a3b':
  changed to use size_t

+++ analysis/modules/ParticleExtractorFromROE/src/ParticleExtractorFromROEModule.cc
@@ -96,7 +96,7 @@ void ParticleExtractorFromROEModule::initialize()
 
 void ParticleExtractorFromROEModule::event()
 {
-  for (int iList = 0; iList < m_outputListNames.size(); iList++) {
+  for (size_t iList = 0; iList < m_outputListNames.size(); iList++) {
 
     const std::string listName = m_outputListNames[iList];
     const std::string antiListName = m_outputAntiListNames[iList];
